+++ src/app/settings/initial-settings.ts
@@ -10,6 +10,9 @@ export interface Settings extends DimApiSettings {
 
   /** Select descriptions to display */
   readonly descriptionsToDisplay: 'bungie' | 'community' | 'both';
+
+  /** Plug the T10 masterwork into D2Y2+ random roll weapons for comparison purposes. */
+  readonly compareWeaponMasterwork: boolean;
 }
 
 export const initialSettingsState: Settings = {
@@ -17,4 +20,5 @@ export const initialSettingsState: Settings = {
   language: defaultLanguage(),
   itemSortReversals: [],
   descriptionsToDisplay: 'both',
+  compareWeaponMasterwork: false,
 };